refactor: MapLetterFacade create 메서드 MapLetterCreateFacade로 분리
