+++ src/main/java/online/bottler/mapletter/domain/policy/MapLetterPolicy.java
@@ -6,4 +6,5 @@ public class MapLetterPolicy {
     }
 
     public static final double VIEW_DISTANCE = 15;
+    public static final int REDIS_SAVED_REPLY = 6;
 }
